⚒️ finish refactoring of moving types into their own package
