+++ packages/@ourworldindata/types/src/GdocsConstants.ts
