+++ packages/@ourworldindata/types/src/dbTypes/Posts.ts
@@ -1,9 +1,9 @@
-import { FormattingOptions } from "@ourworldindata/types"
 import { WP_PostType } from "../wordpress/WordpressTypes.js"
 import {
     OwidArticleBackportingStatistics,
     OwidGdocPostInterface,
 } from "../gdocTypes/Gdoc.js"
+import { FormattingOptions } from "../Wordpress.js"
 
 export interface PostRowPlainFields {
     id: number
