+++ packages/@ourworldindata/types/src/domainTypes/Various.ts
