+++ packages/@ourworldindata/types/src/gdocTypes/Spans.ts
