+++ packages/@ourworldindata/types/src/grapherTypes/GrapherConstants.ts
