+++ packages/@ourworldindata/utils/src/Util.ts
@@ -142,32 +142,248 @@ import { formatLocale, FormatLocaleObject } from "d3-format"
 import striptags from "striptags"
 import parseUrl from "url-parse"
 import {
-    SortOrder,
-    Integer,
-    Time,
-    EPOCH_DATE,
-    ScaleType,
-    VerticalAlign,
-    HorizontalAlign,
+    OwidVariableDisplayConfigInterface,
+    OwidVariableDataTableConfigInteface,
+    OwidChartDimensionInterface,
+    type GitCommit,
+    type Integer,
+    JsonError,
+    type SerializedGridProgram,
+    SiteFooterContext,
+    SuggestedChartRevisionStatus,
+    TaggableType,
+    type TopicId,
+    type OwidVariableId,
+    type RawPageview,
+    type DataValueConfiguration,
+    type DataValueProps,
+    type DataValueQueryArgs,
+    type DataValueResult,
+    type BreadcrumbItem,
+    type KeyInsight,
+    type KeyValueProps,
+    type FormattedPost,
+    type IndexPost,
+    type FullPost,
+    type TocHeading,
+    type TocHeadingWithTitleSupertitle,
+    type Deploy,
+    type DeployChange,
+    DeployStatus,
+    type ChartTag,
+    type ChartTagJoin,
     IDEAL_PLOT_ASPECT_RATIO,
-    GridParameters,
-    OwidGdocPostInterface,
-    OwidGdocJSON,
-    TimeBound,
+    EPOCH_DATE,
+    type Annotation,
+    type Box,
+    type BasicChartInformation,
+    SortBy,
+    type SortConfig,
+    SortOrder,
+    type ValueRange,
+    type Year,
     TimeBoundValue,
-    OwidEnrichedGdocBlock,
-    Span,
+    type TimeRange,
+    type Color,
+    type ColumnSlug,
+    KeyChartLevel,
+    type PrimitiveType,
+    DimensionProperty,
+    type RelatedChart,
+    ToleranceStrategy,
+    ScaleType,
+    Position,
+    type PositionMap,
+    AxisAlign,
+    VerticalAlign,
+    type GridParameters,
+    HorizontalAlign,
+    type EntryMeta,
+    type EntryNode,
+    type PostReference,
+    type CategoryNode,
+    type DocumentNode,
+    type CategoryWithEntries,
+    GraphDocumentType,
+    GraphType,
+    type AlgoliaRecord,
+    type ChartRecord,
+    WP_BlockClass,
+    WP_BlockType,
+    WP_ColumnStyle,
+    WP_PostType,
+    type PostRestApi,
+    type FilterFnPostRestApi,
+    type Ref,
+    type RefDictionary,
+    type BlockPositionChoice,
+    type ChartPositionChoice,
+    type OwidEnrichedGdocBlock,
+    type OwidRawGdocBlock,
+    ChartControlKeyword,
+    ChartTabKeyword,
+    type EnrichedBlockAlign,
+    type RawBlockAlign,
+    type ParseError,
+    BlockImageSize,
+    checkIsBlockImageSize,
+    type RawBlockAllCharts,
+    type RawBlockAdditionalCharts,
+    type RawBlockAside,
+    type RawBlockBlockquote,
+    type RawBlockCallout,
+    type RawBlockChart,
+    type RawBlockChartStory,
+    type RawBlockChartValue,
+    type RawBlockExpandableParagraph,
+    type RawBlockGraySection,
+    type RawBlockHeading,
+    type RawBlockHorizontalRule,
+    type RawBlockHtml,
+    type RawBlockImage,
+    type RawBlockVideo,
+    type RawBlockKeyInsights,
+    type RawBlockList,
+    type RawBlockMissingData,
+    type RawBlockNumberedList,
+    type RawBlockPosition,
+    type RawBlockProminentLink,
+    type RawBlockPullQuote,
+    type RawBlockRecirc,
+    type RawBlockResearchAndWriting,
+    type RawBlockResearchAndWritingLink,
+    type RawBlockScroller,
+    type RawBlockSDGGrid,
+    type RawBlockSDGToc,
+    type RawBlockSideBySideContainer,
+    type RawBlockStickyLeftContainer,
+    type RawBlockStickyRightContainer,
+    type RawBlockText,
+    type RawBlockTopicPageIntro,
+    type RawBlockUrl,
+    tableTemplates,
+    type TableTemplate,
+    tableSizes,
+    type TableSize,
+    type RawBlockTable,
+    type RawBlockTableRow,
+    type RawBlockTableCell,
+    type RawChartStoryValue,
+    type RawRecircLink,
+    type RawSDGGridItem,
+    type RawBlockEntrySummary,
+    type RawBlockEntrySummaryItem,
+    type EnrichedBlockAllCharts,
+    type EnrichedBlockAdditionalCharts,
+    type EnrichedBlockAside,
+    type EnrichedBlockBlockquote,
+    type EnrichedBlockCallout,
+    type EnrichedBlockChart,
+    type EnrichedBlockChartStory,
+    type EnrichedBlockExpandableParagraph,
+    type EnrichedBlockGraySection,
+    type EnrichedBlockHeading,
+    type EnrichedBlockHorizontalRule,
+    type EnrichedBlockHtml,
+    type EnrichedBlockImage,
+    type EnrichedBlockVideo,
+    type EnrichedBlockKeyInsights,
+    type EnrichedBlockKeyInsightsSlide,
+    type EnrichedBlockList,
+    type EnrichedBlockMissingData,
+    type EnrichedBlockNumberedList,
+    type EnrichedBlockProminentLink,
+    type EnrichedBlockPullQuote,
+    type EnrichedBlockRecirc,
+    type EnrichedBlockResearchAndWriting,
+    type EnrichedBlockResearchAndWritingLink,
+    type EnrichedBlockResearchAndWritingRow,
+    type EnrichedBlockScroller,
+    type EnrichedBlockSDGGrid,
+    type EnrichedBlockSDGToc,
+    type EnrichedBlockSideBySideContainer,
+    type EnrichedBlockSimpleText,
+    type EnrichedBlockStickyLeftContainer,
+    type EnrichedBlockStickyRightContainer,
+    type EnrichedBlockText,
+    type EnrichedTopicPageIntroRelatedTopic,
+    type EnrichedTopicPageIntroDownloadButton,
+    type EnrichedBlockTopicPageIntro,
+    type EnrichedChartStoryItem,
+    type EnrichedRecircLink,
+    type EnrichedScrollerItem,
+    type EnrichedSDGGridItem,
+    type EnrichedBlockEntrySummary,
+    type EnrichedBlockEntrySummaryItem,
+    type EnrichedBlockTable,
+    type EnrichedBlockTableRow,
+    type EnrichedBlockTableCell,
+    type RawBlockResearchAndWritingRow,
+    OwidGdocPublicationContext,
+    type OwidGdocErrorMessageProperty,
+    type OwidGdocErrorMessage,
+    OwidGdocErrorMessageType,
+    type OwidGdocLinkJSON,
+    type OwidGdocBaseInterface,
+    type OwidGdocPostContent,
+    type OwidGdocPostInterface,
+    type OwidGdocDataInsightContent,
+    type OwidGdocDataInsightInterface,
+    type MinimalDataInsightInterface,
+    DATA_INSIGHTS_INDEX_PAGE_SIZE,
+    type OwidGdoc,
+    type RawDetail,
     OwidGdocType,
-    EnrichedRecircLink,
-    EnrichedTopicPageIntroRelatedTopic,
-    EnrichedTopicPageIntroDownloadButton,
-    EnrichedScrollerItem,
-    EnrichedBlockKeyInsightsSlide,
+    type OwidGdocStickyNavItem,
+    type OwidGdocJSON,
+    type FaqDictionary,
+    type EnrichedDetail,
+    type EnrichedFaq,
+    type DetailDictionary,
+    GdocsContentSource,
+    type OwidArticleBackportingStatistics,
+    type LinkedChart,
+    DYNAMIC_COLLECTION_PAGE_CONTAINER_ID,
+    DataPageJsonTypeObject,
+    type DataPageJson,
+    type DataPageParseError,
+    type DataPageV2ContentFields,
+    type DataPageDataV2,
+    type DataPageRelatedData,
+    type DataPageRelatedResearch,
+    type PrimaryTopic,
+    type FaqLink,
+    type FaqEntryData,
+    type DisplaySource,
+    type Span,
+    type SpanBold,
+    type SpanDod,
+    type SpanFallback,
+    type SpanItalic,
+    type SpanLink,
+    type SpanNewline,
+    type SpanQuote,
+    type SpanRef,
+    type SpanSimpleText,
+    type SpanSubscript,
+    type SpanSuperscript,
+    type SpanUnderline,
+    type UnformattedSpan,
+    parsePostFormattingOptions,
+    parsePostAuthors,
+    parsePostRow,
+    serializePostRow,
+    parsePostArchieml,
+    type PostRowParsedFields,
+    type PostRowPlainFields,
+    type PostRowUnparsedFields,
+    type PostRowEnriched,
+    type PostRowRaw,
+    type PostRowWithGdocPublishStatus,
     UserCountryInformation,
-    SpanLink,
-    OwidGdocDataInsightInterface,
-    OwidGdoc,
-} from "./owidTypes.js"
+    Time,
+    TimeBound,
+} from "@ourworldindata/types"
 import { PointVector } from "./PointVector.js"
 import React from "react"
 import { match, P } from "ts-pattern"
