+++ packages/@ourworldindata/utils/src/index.ts
@@ -1,290 +1,3 @@
-export type {
-    OwidVariableDisplayConfigInterface,
-    OwidVariableDataTableConfigInteface,
-    OwidChartDimensionInterface,
-} from "./OwidVariableDisplayConfigInterface.js"
-export {
-    type GitCommit,
-    type Integer,
-    JsonError,
-    type SerializedGridProgram,
-    SiteFooterContext,
-    SuggestedChartRevisionStatus,
-    TaggableType,
-    type TopicId,
-    type OwidVariableId,
-    type RawPageview,
-} from "./domainTypes/Various.js"
-export {
-    type DataValueConfiguration,
-    type DataValueProps,
-    type DataValueQueryArgs,
-    type DataValueResult,
-} from "./domainTypes/DataValues.js"
-export {
-    type BreadcrumbItem,
-    type KeyInsight,
-    type KeyValueProps,
-} from "./domainTypes/Site.js"
-export {
-    type FormattedPost,
-    type IndexPost,
-    type FullPost,
-} from "./domainTypes/Posts.js"
-
-export {
-    type TocHeading,
-    type TocHeadingWithTitleSupertitle,
-} from "./domainTypes/Toc.js"
-
-export {
-    type Deploy,
-    type DeployChange,
-    DeployStatus,
-} from "./domainTypes/DeployStatus.js"
-export { type Tag } from "./dbTypes/Tags.js"
-
-export { type ChartTag, type ChartTagJoin } from "./dbTypes/ChartTags.js"
-
-export {
-    IDEAL_PLOT_ASPECT_RATIO,
-    EPOCH_DATE,
-} from "./grapherTypes/GrapherConstants.js"
-
-export {
-    type Annotation,
-    type Box,
-    type BasicChartInformation,
-    SortBy,
-    type SortConfig,
-    SortOrder,
-    type ValueRange,
-    type Year,
-    TimeBoundValue,
-    type TimeRange,
-    type Color,
-    type ColumnSlug,
-    KeyChartLevel,
-    type PrimitiveType,
-    DimensionProperty,
-    type RelatedChart,
-    ToleranceStrategy,
-    ScaleType,
-} from "./grapherTypes/GrapherTypes.js"
-
-export {
-    Position,
-    type PositionMap,
-    AxisAlign,
-    VerticalAlign,
-    type GridParameters,
-    HorizontalAlign,
-} from "./domainTypes/Layout.js"
-
-export {
-    type EntryMeta,
-    type EntryNode,
-    type PostReference,
-    type CategoryNode,
-    type DocumentNode,
-    type CategoryWithEntries,
-    GraphDocumentType,
-    GraphType,
-    type AlgoliaRecord,
-    type ChartRecord,
-} from "./domainTypes/ContentGraph.js"
-export {
-    WP_BlockClass,
-    WP_BlockType,
-    WP_ColumnStyle,
-    WP_PostType,
-    type PostRestApi,
-    type FilterFnPostRestApi,
-} from "./wordpress/WordpressTypes.js"
-
-export {
-    type Ref,
-    type RefDictionary,
-    type BlockPositionChoice,
-    type ChartPositionChoice,
-    type OwidEnrichedGdocBlock,
-    type OwidRawGdocBlock,
-    ChartControlKeyword,
-    ChartTabKeyword,
-    type EnrichedBlockAlign,
-    type RawBlockAlign,
-    type ParseError,
-    BlockImageSize,
-    checkIsBlockImageSize,
-    type RawBlockAllCharts,
-    type RawBlockAdditionalCharts,
-    type RawBlockAside,
-    type RawBlockBlockquote,
-    type RawBlockCallout,
-    type RawBlockChart,
-    type RawBlockChartStory,
-    type RawBlockChartValue,
-    type RawBlockExpandableParagraph,
-    type RawBlockGraySection,
-    type RawBlockHeading,
-    type RawBlockHorizontalRule,
-    type RawBlockHtml,
-    type RawBlockImage,
-    type RawBlockVideo,
-    type RawBlockKeyInsights,
-    type RawBlockList,
-    type RawBlockMissingData,
-    type RawBlockNumberedList,
-    type RawBlockPosition,
-    type RawBlockProminentLink,
-    type RawBlockPullQuote,
-    type RawBlockRecirc,
-    type RawBlockResearchAndWriting,
-    type RawBlockResearchAndWritingLink,
-    type RawBlockScroller,
-    type RawBlockSDGGrid,
-    type RawBlockSDGToc,
-    type RawBlockSideBySideContainer,
-    type RawBlockStickyLeftContainer,
-    type RawBlockStickyRightContainer,
-    type RawBlockText,
-    type RawBlockTopicPageIntro,
-    type RawBlockUrl,
-    tableTemplates,
-    type TableTemplate,
-    tableSizes,
-    type TableSize,
-    type RawBlockTable,
-    type RawBlockTableRow,
-    type RawBlockTableCell,
-    type RawChartStoryValue,
-    type RawRecircLink,
-    type RawSDGGridItem,
-    type RawBlockEntrySummary,
-    type RawBlockEntrySummaryItem,
-    type EnrichedBlockAllCharts,
-    type EnrichedBlockAdditionalCharts,
-    type EnrichedBlockAside,
-    type EnrichedBlockBlockquote,
-    type EnrichedBlockCallout,
-    type EnrichedBlockChart,
-    type EnrichedBlockChartStory,
-    type EnrichedBlockExpandableParagraph,
-    type EnrichedBlockGraySection,
-    type EnrichedBlockHeading,
-    type EnrichedBlockHorizontalRule,
-    type EnrichedBlockHtml,
-    type EnrichedBlockImage,
-    type EnrichedBlockVideo,
-    type EnrichedBlockKeyInsights,
-    type EnrichedBlockKeyInsightsSlide,
-    type EnrichedBlockList,
-    type EnrichedBlockMissingData,
-    type EnrichedBlockNumberedList,
-    type EnrichedBlockProminentLink,
-    type EnrichedBlockPullQuote,
-    type EnrichedBlockRecirc,
-    type EnrichedBlockResearchAndWriting,
-    type EnrichedBlockResearchAndWritingLink,
-    type EnrichedBlockResearchAndWritingRow,
-    type EnrichedBlockScroller,
-    type EnrichedBlockSDGGrid,
-    type EnrichedBlockSDGToc,
-    type EnrichedBlockSideBySideContainer,
-    type EnrichedBlockSimpleText,
-    type EnrichedBlockStickyLeftContainer,
-    type EnrichedBlockStickyRightContainer,
-    type EnrichedBlockText,
-    type EnrichedTopicPageIntroRelatedTopic,
-    type EnrichedTopicPageIntroDownloadButton,
-    type EnrichedBlockTopicPageIntro,
-    type EnrichedChartStoryItem,
-    type EnrichedRecircLink,
-    type EnrichedScrollerItem,
-    type EnrichedSDGGridItem,
-    type EnrichedBlockEntrySummary,
-    type EnrichedBlockEntrySummaryItem,
-    type EnrichedBlockTable,
-    type EnrichedBlockTableRow,
-    type EnrichedBlockTableCell,
-    type RawBlockResearchAndWritingRow,
-} from "./gdocTypes/ArchieMlComponents.js"
-export {
-    OwidGdocPublicationContext,
-    type OwidGdocErrorMessageProperty,
-    type OwidGdocErrorMessage,
-    OwidGdocErrorMessageType,
-    type OwidGdocLinkJSON,
-    type OwidGdocBaseInterface,
-    type OwidGdocPostContent,
-    type OwidGdocPostInterface,
-    type OwidGdocDataInsightContent,
-    type OwidGdocDataInsightInterface,
-    type MinimalDataInsightInterface,
-    DATA_INSIGHTS_INDEX_PAGE_SIZE,
-    type OwidGdoc,
-    type RawDetail,
-    OwidGdocType,
-    type OwidGdocStickyNavItem,
-    type OwidGdocJSON,
-    type FaqDictionary,
-    type EnrichedDetail,
-    type EnrichedFaq,
-    type DetailDictionary,
-    GdocsContentSource,
-    type OwidArticleBackportingStatistics,
-    type LinkedChart,
-    DYNAMIC_COLLECTION_PAGE_CONTAINER_ID,
-} from "./gdocTypes/Gdoc.js"
-
-export {
-    DataPageJsonTypeObject,
-    type DataPageJson,
-    type DataPageParseError,
-    type DataPageV2ContentFields,
-    type DataPageDataV2,
-    type DataPageRelatedData,
-    type DataPageRelatedResearch,
-    type PrimaryTopic,
-    type FaqLink,
-    type FaqEntryData,
-    type DisplaySource,
-} from "./gdocTypes/Datapage.js"
-
-export {
-    type Span,
-    type SpanBold,
-    type SpanDod,
-    type SpanFallback,
-    type SpanItalic,
-    type SpanLink,
-    type SpanNewline,
-    type SpanQuote,
-    type SpanRef,
-    type SpanSimpleText,
-    type SpanSubscript,
-    type SpanSuperscript,
-    type SpanUnderline,
-    type UnformattedSpan,
-} from "./gdocTypes/Spans.js"
-
-export {
-    parsePostFormattingOptions,
-    parsePostAuthors,
-    parsePostRow,
-    serializePostRow,
-    parsePostArchieml,
-} from "./dbTypes/PostsUtilities.js"
-
-export {
-    type PostRowParsedFields,
-    type PostRowPlainFields,
-    type PostRowUnparsedFields,
-    type PostRowEnriched,
-    type PostRowRaw,
-    type PostRowWithGdocPublishStatus,
-} from "./dbTypes/Posts.js"
-
 export {
     pairs,
     type NoUndefinedValues,
@@ -501,8 +214,6 @@ export type {
     utcType,
 } from "./dayjs.js"
 
-export type { OwidSource } from "./OwidSource.js"
-export type { OwidOrigin } from "./OwidOrigin.js"
 export {
     formatValue,
     checkIsVeryShortUnit,
@@ -552,26 +263,7 @@ export {
 
 export { PointVector } from "./PointVector.js"
 
-export {
-    OwidVariableDisplayConfig,
-    type OwidVariableWithSource,
-    type OwidVariableWithSourceAndDimension,
-    type OwidVariableWithSourceAndDimensionWithoutId,
-    type OwidVariableMixedData,
-    type OwidVariableWithDataAndSource,
-    type OwidVariableDimension,
-    type OwidVariableDimensions,
-    type OwidVariableDataMetadataDimensions,
-    type MultipleOwidVariableDataDimensionsMap,
-    type OwidVariableDimensionValuePartial,
-    type OwidVariableDimensionValueFull,
-    type OwidVariablePresentation,
-    type OwidEntityKey,
-    type OwidLicense,
-    type OwidProcessingLevel,
-    type IndicatorTitleWithFragments,
-    joinTitleFragments,
-} from "./OwidVariable.js"
+export { OwidVariableDisplayConfig } from "./OwidVariable.js"
 
 export {
     type QueryParams,
@@ -674,8 +366,6 @@ export {
     generateSrcSet,
     getFilenameWithoutExtension,
     getFilenameAsPng,
-    type GDriveImageMetadata,
-    type ImageMetadata,
     type SourceProps,
     generateSourceProps,
     getFeaturedImageFilename,
@@ -685,22 +375,15 @@ export {
 export { Tippy, TippyIfInteractive } from "./Tippy.js"
 
 export {
-    ALL_CHARTS_ID,
-    LICENSE_ID,
-    CITATION_ID,
-    ENDNOTES_ID,
-    KEY_INSIGHTS_ID,
-    RESEARCH_AND_WRITING_ID,
-    IMAGES_DIRECTORY,
-    gdocUrlRegex,
-    gdocIdRegex,
-} from "./GdocsConstants.js"
-
-export {
     extractFormattingOptions,
     parseFormattingOptions,
     parseKeyValueArgs,
 } from "./wordpressUtils.js"
+
+// This re-exports everything in the types package from the utils package. This is done so that
+// the transition is easier - we might want to get rid of this and rewrite all the imports instead
+// but it's a lot of work
+export * from "@ourworldindata/types"
 
 export {
     getErrorMessageDonation,
